# swagger/openapi.json
{
  "components": {
    "schemas": {
      "ProjectDto": {
        "type": "object",
        "required": [
          "name",
          "organization"
        ],
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationDto"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UserInOrganizationDto": {
        "type": "object",
        "required": [
          "organizationRole",
          "username"
        ],
        "properties": {
          "organizationRole": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ]
          },
          "username": {
            "type": "string"
          }
        }
      },
      "ProjectInOrganizationDto": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "OrganizationDetailsDto": {
        "type": "object",
        "required": [
          "organization",
          "projects",
          "users"
        ],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectInOrganizationDto"
            }
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationDto"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserInOrganizationDto"
            }
          }
        }
      },
      "UserDto": {
        "type": "object",
        "required": [
          "email",
          "name",
          "username"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "ProjectCreationDto": {
        "type": "object",
        "required": [
          "projectName"
        ],
        "properties": {
          "projectName": {
            "type": "string"
          }
        }
      },
      "AddUserToOrganizationDto": {
        "type": "object",
        "required": [
          "role",
          "username"
        ],
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ]
          },
          "username": {
            "type": "string"
          }
        }
      },
      "OrganizationDto": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
    "/organizations": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationDto"
              }
            }
          },
          "required": true
        },
        "operationId": "create_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        },
        "tags": [
          "organizations-controller"
        ]
      },
      "get": {
        "operationId": "list_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "organizations-controller"
        ]
      }
    },
    "/organizations/{organizationName}/projects": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreationDto"
              }
            }
          },
          "required": true
        },
        "operationId": "create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationName",
            "required": true
          }
        ],
        "tags": [
          "project-controller"
        ]
      },
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationName",
            "required": true
          }
        ],
        "tags": [
          "project-controller"
        ]
      }
    },
    "/users": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "operationId": "create_2",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": [
          "user-controller"
        ]
      },
      "get": {
        "operationId": "list_2",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "user-controller"
        ]
      }
    },
    "/organizations/{organizationName}/projects/{projectName}": {
      "get": {
        "operationId": "get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationName",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectName",
            "required": true
          }
        ],
        "tags": [
          "project-controller"
        ]
      },
      "delete": {
        "operationId": "delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationName",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectName",
            "required": true
          }
        ],
        "tags": [
          "project-controller"
        ]
      }
    },
    "/organizations/{organizationName}/users": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserToOrganizationDto"
              }
            }
          },
          "required": true
        },
        "operationId": "addUser",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationName",
            "required": true
          }
        ],
        "tags": [
          "organizations-controller"
        ]
      }
    },
    "/users/{username}": {
      "get": {
        "operationId": "get_2",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "username",
            "required": true
          }
        ],
        "tags": [
          "user-controller"
        ]
      },
      "delete": {
        "operationId": "delete_2",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "username",
            "required": true
          }
        ],
        "tags": [
          "user-controller"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "operationId": "update_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "username",
            "required": true
          }
        ],
        "tags": [
          "user-controller"
        ]
      }
    },
    "/organizations/{name}": {
      "get": {
        "operationId": "get_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetailsDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "tags": [
          "organizations-controller"
        ]
      },
      "delete": {
        "operationId": "delete_1",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "tags": [
          "organizations-controller"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationDto"
              }
            }
          },
          "required": true
        },
        "operationId": "update",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "name",
            "required": true
          }
        ],
        "tags": [
          "organizations-controller"
        ]
      }
    },
    "/organizations/{organizationName}/users/{username}": {
      "delete": {
        "operationId": "removeUser",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationName",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "username",
            "required": true
          }
        ],
        "tags": [
          "organizations-controller"
        ]
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}